PhlyTest namespace in autoload-dev
Hello. I want to include this project in [my list of lightweight libraries](https://github.com/jm42/lightweight-php), but it has something that I find problematic to calculate the "weight" of this projects.

The test namespace is included in the autoload configuration, always. See, I use the directories in the autoload configuration of composer to know which are the *source* folders to parse and calculate the "weight".

Since I think the test namespace will be only used when the dev requirements have been installed it seems like a plausible change.