Refactor for immutabile messages
Refactored the MessageTrait, Request, Response, and ServerRequest to be
immutable; all setters clone and return the clone with the updated values.

Because of this change, the ServerRequestFactory needed updates to ensure it was
building and returning the request instance properly.

The Server::listen() method needed to be altered to check the return value of
the callable; if a response is returned, that response is used to pass to
send(), but if none is returned, the composed response is used. This allows the
callable to "alter" the response.