Updated Uri to validate scheme when constructed.
Changed parseUri() to filter the scheme so a Uri cannot be constructed with an invalid scheme. The list of schemes can be modified by an extending class if an implementation wanted to support additional schemes.

Also altered getPort() to return null if the port was set to the standard port for the scheme.